+++ system/src/Grav/Framework/Form/FormFlash.php
@@ -27,7 +27,9 @@ class FormFlash implements \JsonSerializable
     /** @var array */
     protected $user;
     /** @var int */
-    protected $timestamp;
+    protected $createdTimestamp;
+    /** @var int */
+    protected $updatedTimestamp;
     /** @var array */
     protected $data;
     /** @var array */
@@ -71,13 +73,14 @@ class FormFlash implements \JsonSerializable
             $this->formName = $content['form'] ?? $formName;
             $this->url = $data['url'] ?? '';
             $this->user = $data['user'] ?? null;
-            $this->timestamp = $data['timestamp'] ?? time();
+            $this->updatedTimestamp = $data['timestamps']['updated'] ?? time();
+            $this->createdTimestamp = $data['timestamps']['created'] ?? $this->updatedTimestamp;
             $this->data = $data['data'] ?? null;
             $this->files = $data['files'] ?? [];
         } else {
             $this->formName = $formName;
             $this->url = '';
-            $this->timestamp = time();
+            $this->createdTimestamp = $this->updatedTimestamp = time();
             $this->files = [];
         }
     }
@@ -101,9 +104,17 @@ class FormFlash implements \JsonSerializable
     /**
      * @return int
      */
-    public function getTimestamp(): int
+    public function getCreatedTimestamp(): int
     {
-        return $this->timestamp;
+        return $this->createdTimestamp;
+    }
+
+    /**
+     * @return int
+     */
+    public function getUpdatedTimestamp(): int
+    {
+        return $this->updatedTimestamp;
     }
 
     /**
@@ -119,9 +130,15 @@ class FormFlash implements \JsonSerializable
      */
     public function save(): self
     {
-        $file = $this->getTmpIndex();
-        $file->save($this->jsonSerialize());
-        $this->exists = true;
+        if ($this->data || $this->files) {
+            // Only save if there is data or files to be saved.
+            $file = $this->getTmpIndex();
+            $file->save($this->jsonSerialize());
+            $this->exists = true;
+        } elseif ($this->exists) {
+            // Delete empty form flash if it exists (it carries no information).
+            return $this->delete();
+        }
 
         return $this;
     }
@@ -353,7 +370,10 @@ class FormFlash implements \JsonSerializable
             'unique_id' => $this->uniqueId,
             'url' => $this->url,
             'user' => $this->user,
-            'timestamp' => time(),
+            'timestamps' => [
+                'created' => $this->createdTimestamp,
+                'updated' => time(),
+            ],
             'data' => $this->data,
             'files' => $this->files
         ];