+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 1. [](#new)
     * Added `FormTrait::getAllFlashes()` method to get all the available form flash objects for the form
-    * Added timestamp to `FormFlash` objects
+    * Added creation and update timestamps to `FormFlash` objects
 1. [](#bugfix)
     * Fixed error in `ImageMedium::url()` if the image cache folder does not exist
     * Fixed empty form flash name after file upload or form state update